Rename and refactor fisher_vector file
