+++ skimage/feature/meson.build
@@ -63,7 +63,7 @@ python_sources = [
   'brief.py',
   'censure.py',
   'corner.py',
-  'fisher_vector.py',
+  '_fisher_vector.py',
   'haar.py',
   'match.py',
   'orb.py',
